Repurpose dotpkg.resolve for manifest resolution
